@@ -7,9 +7,9 @@
     "@testing-library/react": "^12.0.0",
     "@testing-library/user-event": "^12.0.0",
     "alchemy-sdk": "^2.2.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.21.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
